refactor: showErrorAlert location -> extension UIViewController
